+++ deepinv/distributed/__init__.py
@@ -1,11 +1,11 @@
-from .distrib_framework import (
+from deepinv.distributed.distrib_framework import (
     DistributedContext,
     DistributedPhysics,
     DistributedLinearPhysics,
     DistributedProcessing,
     DistributedDataFidelity,
 )
-from .distribute import distribute
+from deepinv.distributed.distribute import distribute
 
 __all__ = [
     "DistributedContext",