+++ docs/source/conf.py
@@ -338,7 +338,7 @@ sphinx_gallery_conf = {
             "../../examples/self-supervised-learning",
             "../../examples/adversarial-learning",
             "../../examples/external-libraries",
-            "../../examples/distrib",
+            "../../examples/distributed",
         ]
     ),
     "within_subsection_order": MySortKey,